+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v2.1.0 - 2022-05-22
+
+### Added
+
+- `GenerateCore()` and `LastStatus()` to Generator
+
+### Maintenance
+
+- Updated README
+
 ## v2.0.0 - 2022-05-01
 
 ### Changed